# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EventSphere API - College Events Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{eventId}/registrations:
    get:
      operationId: getEventRegistrations
      tags: [registrations]
      summary: Get registrations for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Registrations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
  /events/{eventId}/register:
    post:
      operationId: registerForEvent
      tags: [registrations]
      summary: Register for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
    delete:
      operationId: cancelRegistration
      tags: [registrations]
      summary: Cancel registration for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cancelled
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
            enum: [participant, organizer, admin]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended]
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
...
  /users/{userId}/status:
    put:
      operationId: updateUserStatus
      tags: [users]
      summary: Update user status (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserStatusBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /auth/signup:
    post:
      operationId: signUp
      tags: [users]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [users]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
components:
  schemas:
...
    Registration:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        eventTitle:
          type: string
        userId:
          type: integer
        userName:
          type: string
        userEmail:
          type: string
        ticketType:
          type: string
          enum: [general, vip]
        status:
          type: string
          enum: [registered, waitlist, checked_in, cancelled]
        passCode:
          type: string
        registeredAt:
          type: string
      required: [id, eventId, eventTitle, userId, userName, userEmail, ticketType, status, registeredAt]
    RegisterBody:
      type: object
      properties:
        ticketType:
          type: string
          enum: [general, vip]
        userId:
          type: integer
      required: [ticketType, userId]
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [participant, organizer, admin]
        status:
          type: string
          enum: [active, suspended]
        department:
          type: string
        year:
          type: integer
        avatarUrl:
          type: string
        joinedAt:
          type: string
      required: [id, fullName, email, role, status, joinedAt]
    UpdateUserStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, suspended]
      required: [status]
    SignUpBody:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [participant, organizer]
      required: [fullName, email, password, role]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
        message:
          type: string
      required: [user, token, message]
...